+++ openfisca_aotearoa/variables/demographics/ages.py
@@ -47,7 +47,7 @@ class age(Variable):
 class age_of_youngest(Variable):
     value_type = int
     entity = Family
-    definition_period = MONTH
+    definition_period = DAY
     unit = 'years'
     label = u"The age of the youngest member of a family"
     # A person's age is computed according to their birth date.
@@ -59,7 +59,7 @@ class age_of_youngest(Variable):
 class age_of_partner(Variable):
     value_type = int
     entity = Person
-    definition_period = MONTH
+    definition_period = DAY
     unit = 'years'
     label = u"The age of partner in a family"